### 2019-07-13 #1567 @whedon
[PRE REVIEW]: lfda: Local Fisher Discriminant Analysis in R
**Submitting author:** @terrytangyuan (<a href="http://orcid.org/0000-0001-5243-233X">Yuan Tang</a>)
**Repository:** <a href="https://github.com/terrytangyuan/lfda/" target ="_blank">https://github.com/terrytangyuan/lfda/</a>
**Version:** v1.1.2
