### 2019-07-16 #1567 @sctyner
Looks like you don't need me, but just in case, I'm happy to review.
