### 2019-07-16 #1567 @terrytangyuan
Looks like @sctyner is the only one who didn't mention any availability so I guess she should be able to review this quicker than anyone else. @corybrunson reviewed for my `autoplotly` JOSS paper before so he should be familiar with some of the related content in this paper. But I'll leave this to our editor @arfon to decide on the reviewers. 👋
