### 2019-07-23 terrytangyuan/lfda#22 @sctyner
JOSS Review General Checks
This repository does not pass two of the general checks from the JOSS [review](https://github.com/openjournals/joss-reviews/issues/1572):

1. "License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?" No. The current [license](https://github.com/terrytangyuan/lfda/blob/master/LICENSE) is not adequate.
2. "Version: Does the release version given match the GitHub release (v1.1.2)?" The [description](https://github.com/terrytangyuan/lfda/blob/master/DESCRIPTION) file lists the current version as 1.1.3.
