### 2019-07-23 #1572 @sctyner
The repository does not pass the License and Version general checks. Issue [here](https://github.com/terrytangyuan/lfda/issues/22).
