### 2019-07-23 #1572 @sctyner
@arfon I have a question about authorship. The DESCRIPTION file of the R package list the Authors as, "Yuan Tang. Contributions from Wenxuan Li, Nan Xiao and Zachary Deane-Mayer." However, only Yuan Tang and Wenxuan Li are listed as authors in the [paper.md](https://github.com/terrytangyuan/lfda/blob/master/paper/paper.md) file. Does one or both of these need to change in order to be acceptable? Based on my reading of the submission requirements and checklist above, it seems that the authors on the paper should include all four names given in the description file. What is correct? Thanks in advance for your guidance.
