### 2019-07-23 #1572 @terrytangyuan
@sctyner I've fixed the issue with the license. Regarding the other issue with the version, I was hoping to release v1.1.3 on Github and on CRAN after the paper got accepted so I could include the full version of the paper as part of the release. @arfon Could you double check to see if this is the right process?

Regarding the author names, I asked Nan before and he preferred not to be included since he did not contribute much. Zachary helped with the CI but never heard of him afterwards (I sent an email earlier to express my willingness to include him as co-author but he hasn't responded for years).
