### 2019-07-23 terrytangyuan/lfda#23 @sctyner
JOSS Review Tests
Again for the JOSS [review](https://github.com/openjournals/joss-reviews/issues/1572). There are six empty tests (i.e. they are not testing anything):

- In [test_helper_functions.R](https://github.com/terrytangyuan/lfda/blob/master/tests/testthat/test_helper_functions.R)
    - "kmatrixGauss function works"
    - "repmat works"
- In [test_klfda.R](https://github.com/terrytangyuan/lfda/blob/master/tests/testthat/test_klfda.R)
    - "klfda works"
- In [test_visualization.R](https://github.com/terrytangyuan/lfda/blob/master/tests/testthat/test_visualization.R)
    - "klfda visualization works"
    - "lfda visualization works"
    - "self visualization works"
