### 2019-07-23 #1572 @sctyner
There are six empty tests. Issue [here](https://github.com/terrytangyuan/lfda/issues/23)
