### 2019-07-23 terrytangyuan/lfda#23 @terrytangyuan
What do you mean? They are testing the exposed functions from the package.
