### 2019-07-23 #1572 @terrytangyuan
@sctyner Could you elaborate on what you meant by “empty tests”? Those are testing the exposed core functions from the package, e.g. lfda(), klfda(), etc.
