### 2019-07-23 terrytangyuan/lfda#23 @sctyner
They do not contain any expectations. From the `test_that()` [documentation](https://testthat.r-lib.org/reference/test_that.html), there are 2 arguments to `test_that()`: `desc` and `code`. `desc` is the "test name. Names should be kept as brief as possible, as they are often used as line prefixes." `code` is "test code containing expectations."

Here is the output from `devtools::test()`:

```r
> devtools::test()
Loading lfda
Testing lfda
✔ |  OK F W S | Context
✔ |   1     2 | helper functions
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test_helper_functions.R:5: skip: kmatrixGauss function works
Empty test

test_helper_functions.R:9: skip: repmat works
Empty test
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |   0     1 | klfda
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test_klfda.R:5: skip: klfda works
Empty test
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |   3       | lfda
✔ |   3       | self [0.4 s]
✔ |   5     3 | visualization [0.2 s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
test_visualization.R:5: skip: klfda visualization works
Empty test

test_visualization.R:16: skip: lfda visualization works
Empty test

test_visualization.R:27: skip: self visualization works
Empty test
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 0.7 s

OK:       12
Failed:   0
Warnings: 0
Skipped:  6
```
