### 2019-07-23 terrytangyuan/lfda#24 @sctyner
JOSS Review: Functionality
For JOSS [review](https://github.com/openjournals/joss-reviews/issues/1572). I have not been able to plot any of the example models in the README using the `plot(model, y)` syntax. The code runs and no errors are produced, and the `model` object looks ok, but `plot(model, y)` does not produce anything.

What could I be missing? I have all the dependencies and suggests installed.
