+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: lfda
 Type: Package
 Title: Local Fisher Discriminant Analysis
-Version: 1.1.3
+Version: 1.1.2
 Date: 2019-07-13
 Author: Yuan Tang. Contributions from Wenxuan Li, Nan Xiao and Zachary Deane-Mayer.
 URL: https://github.com/terrytangyuan/lfda
