### 2019-07-23 #1572 @terrytangyuan
@sctyner Thanks. Just saw your reply on the issue. BTW, I changed the version back to v1.1.2 in DESCRIPTION so this should no longer be an issue.
