### 2019-07-23 terrytangyuan/lfda#24 @sctyner
Here is my session info. Should have included it at first!
```r
> sessionInfo()
R version 3.5.3 (2019-03-11)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS Mojave 10.14.6

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.5/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
[1] rARPACK_0.11-0  ggfortify_0.4.7 ggplot2_3.2.0   lfda_1.1.3      rgl_0.100.19    testthat_2.1.1

 [1] Rcpp_1.0.1              lattice_0.20-38         tidyr_0.8.3             prettyunits_1.0.2       ps_1.3.0
 [6] assertthat_0.2.1        rprojroot_1.3-2         digest_0.6.20           packrat_0.5.0           RSpectra_0.15-0
[11] mime_0.7                R6_2.4.0                plyr_1.8.4              backports_1.1.4         evaluate_0.13
[16] pillar_1.4.2            rlang_0.4.0             lazyeval_0.2.2          rstudioapi_0.10         miniUI_0.1.1.1
[21] callr_3.2.0             Matrix_1.2-15           DT_0.5                  rmarkdown_1.11          labeling_0.3
[26] desc_1.2.0              devtools_2.1.0          webshot_0.5.1           stringr_1.4.0           htmlwidgets_1.3
[31] munsell_0.5.0           shiny_1.3.2             compiler_3.5.3          httpuv_1.5.1            xfun_0.7.2
[36] pkgconfig_2.0.2         pkgbuild_1.0.3          htmltools_0.3.6         tidyselect_0.2.5        gridExtra_2.3
[41] tibble_2.1.3            crayon_1.3.4            dplyr_0.8.3.9000        withr_2.1.2             later_0.8.0
[46] grid_3.5.3              jsonlite_1.6            xtable_1.8-4            gtable_0.3.0            magrittr_1.5
[51] covr_3.2.1              scales_1.0.0            stringi_1.4.3           cli_1.1.0               fs_1.2.7
[56] promises_1.0.1          remotes_2.1.0           rex_1.1.2               tools_3.5.3             manipulateWidget_0.10.0
[61] glue_1.3.1              purrr_0.3.2             crosstalk_1.0.0         processx_3.3.0          pkgload_1.0.2
[66] yaml_2.2.0              colorspace_1.4-1        sessioninfo_1.1.1       memoise_1.1.0           knitr_1.22.10
[71] usethis_1.5.0
```
