### 2019-07-23 terrytangyuan/lfda#24 @terrytangyuan
@sctyner Thanks. Could you also try this visualization demo using ggfortify (based on ggplot2)? https://cran.r-project.org/web/packages/ggfortify/vignettes/plot_pca.html
It seems like fine on CRAN. `plot()` is based on `rgl` which might be tricky on different platforms.
