### 2019-07-23 terrytangyuan/lfda#24 @terrytangyuan
Also, maybe try running `rgl` directly, e.g. [this tutorial](https://cran.r-project.org/web/packages/rgl/vignettes/rgl.html) so that we can isolate the issue? I just tried it locally on MacOS and everything works fine.
