### 2019-07-23 terrytangyuan/lfda#24 @sctyner
I ran through the 2D examples just fine! I have `rgl` installed, but I did just do some updates on my MacBook today, so maybe that's the issue? I'll have to look into that.
