### 2019-07-23 terrytangyuan/lfda#24 @sctyner
For the tutorial you linked, neither of the first two examples work for me. This must be an update issue.
