### 2019-07-23 terrytangyuan/lfda#24 @sctyner
Ok, I re-installed `rgl` and `lfda`. The `plot(model,y)` statement produces a window, which it wasn't before, so that's encouraging! The window can't be resized though. This is the same for the examples you linked. Since I just updated this morning, I'll assume that it's the Xcode update's issue. :)
