### 2019-07-23 terrytangyuan/lfda#24 @terrytangyuan
@sctyner Yea that seems like an issue with `rgl` on your Mac. Should we close this now as it is not related to this package?
