### 2019-07-23 bde69f9
More tests
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
