+++ tests/testthat/test_klfda.R
@@ -1,12 +1,20 @@
-context('klfda')
+context('Test klfda functionalities')
 
 data(iris)
 
-test_that('klfda works', {
+test_that('klfda works correctly for different metrics on the data transformed by kmatrixGauss()', {
   k <- kmatrixGauss(iris[,-5])
   y <- iris[,5]
   r <- 3
-  klfda(k,y,r,metric="plain")
-  klfda(k,y,r,metric="weighted")
-  klfda(k,y,r,metric="orthonormalized")
+  result <- klfda(k, y, r, metric = "plain")
+  expect_equal(dim(result$T), c(150, 3))
+  expect_equal(dim(result$Z), c(150, 3))
+  
+  result <- klfda(k, y, r, metric = "weighted")
+  expect_equal(dim(result$T), c(150, 3))
+  expect_equal(dim(result$Z), c(150, 3))
+  
+  result <- klfda(k, y, r, metric = "orthonormalized")
+  expect_equal(dim(result$T), c(150, 3))
+  expect_equal(dim(result$Z), c(150, 3))
 })
