### 2019-07-23 #1572 @sctyner
Ok, I am pretty much done. The automated tests and hearing from @arfon about the versioning are the only unresolved things.
