### 2019-07-23 b574e7d
Add tests for visualization
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
