### 2019-07-23 terrytangyuan/lfda#23 @terrytangyuan
This has been fixed:
```
> devtools::test(".")
Loading lfda
Testing lfda
✔ | OK F W S | Context
✔ |  4       | Test the helper functions used in lfda implementations
✔ |  6       | Test klfda functionalities [0.1 s]
✔ | 13       | Test lfda functionalities [0.1 s]
✔ |  9       | self [0.5 s]
✔ |  8       | Test visualization on different lfda models [0.2 s]

══ Results ════════════════════════════════════════════════════════════════════════════════
Duration: 1.1 s

OK:       40
Failed:   0
Warnings: 0
Skipped:  0
```
