### 2019-07-23 #1572 @terrytangyuan
@sctyner Just fixed the issue with tests. Regarding the versioning, I fixed the version number in `DESCRIPTION` file so I think it should be fine now. Please let me know if there's anything else I should resolve. Thank you!
