### 2019-07-24 terrytangyuan/lfda#23 @sctyner
Just so I'm clear - the point of testing the plot method is to make sure it runs without error. Is that correct?
