### 2019-07-24 terrytangyuan/lfda#23 @terrytangyuan
@sctyner Yes, I am not sure if there's a good way to test the rendered plots as far as I know. I think as long as the result from `lfda()` is correct, the rest should be straightforward usage of `rgl`.
