### 2019-07-24 terrytangyuan/lfda#23 @sctyner
Ok. I know that [`vdiffr`](https://github.com/r-lib/vdiffr) is available to test the appearance of plots, but I've never seen it applied to 3D plots.
