### 2019-07-24 #1572 @sctyner
My review is complete.
