### 2019-07-24 terrytangyuan/lfda#25 @strengejacke
[JOSS review] - missing DOIs in paper
Related to https://github.com/openjournals/joss-reviews/issues/1572

Please add a DOI to all references where available. For example, Scholkopft/Mullert or the R Journal have DOIs.
