### 2019-07-24 terrytangyuan/lfda#26 @strengejacke
[JOSS review] - paper, statement of need
Related to https://github.com/openjournals/joss-reviews/issues/1572

The paper clearly states what the package does, but does not emphasize why it's needed. Please explain if the functionality in the lfda package is not provided by other packages, or what other needs are being addressed by your package.
