### 2019-07-24 terrytangyuan/lfda#28 @strengejacke
[JOSS review] - provide working examples
related to https://github.com/openjournals/joss-reviews/issues/1572

I can't quickly check if the package is working, because examples don't run. See, e.g., `?lfda`. The provided examples does not run. Please check your docs / examples and provide running examples for the user.
