### 2019-07-24 #1572 @strengejacke
I have opened several issues (https://github.com/terrytangyuan/lfda/issues) that need to be addressed, some of them are required to complete my review.

I think the submission is well prepared, so I would consider the issues as minor changes.
