### 2019-07-24 terrytangyuan/lfda#28 @sctyner
@strengejacke were you able to get the plots on the README to render? I had issues which I assumed were due to my recent MacOS update, but this issue is making me reconsider.
