### 2019-07-24 terrytangyuan/lfda#28 @terrytangyuan
I think this issue is only related to the examples in the roxygen docs where they are marked as `\dontrun` which is unintended. If so, I can fix them soon. The examples should be running fine.
