### 2019-07-24 terrytangyuan/lfda#28 @strengejacke
I copy/pasted the example into the R console, so the code was executed, but failed.
