### 2019-07-24 terrytangyuan/lfda#28 @terrytangyuan
Yes, since "trainData" variable does not exist yet in the examples. The examples were not meant to run successfully. It's only instructional and users are responsible for providing the data. Now that I re-think about it. I can for sure replace them with some running examples there.
