### 2019-07-24 terrytangyuan/lfda#28 @strengejacke
> were you able to get the plots on the README to render?

Yes, that worked for me. It seems it requires RGL, which might cause issues on OSX. Maybe it's possible to replace with a plot-method that works on all OS?
