### 2019-07-24 terrytangyuan/lfda#28 @terrytangyuan
Well `rgl` works for me on both MacOS and Windows. I really think that the `rgl` issue is not related to this package. Any package would possibly have issues with some versions of OS.
