### 2019-07-24 af280cc
Fixes #28: Fix examples in roxygen comments
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
