### 2019-07-24 3f518b3
Fixes #29: Rephrase some terms in the paper
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
