+++ paper/paper.md
@@ -29,11 +29,11 @@ affiliations:
 
 # Summary
 
-*Fisher discriminant analysis* [@fdapaper] is a popular choice to reduce the dimensionality of the original data set. It maximizes between-class scatter and minimizes within-class scatter. It works really well in practice but lacks some considerations for multimodality. Multimodality exists within many applications, such as disease diagnosis, where there may be multiple causes for a particular disease. In this situation, *Fisher discriminant analysis* cannot capture the multimodal characteristics of the clusters. To deal with multimodality, *local-preserving projection* [@lpppaper] preserves the local structure of the data in that it keeps nearby data pairs in the original data space close in the embedding space. As a result, multimodal data could be embedded and its local structure will not be lost.
+*Fisher discriminant analysis* [@fdapaper] is a popular choice to reduce the dimensionality of the original dataset. It maximizes between-class scatter and minimizes within-class scatter. It works really well in practice but lacks some considerations for multimodality. Multimodality exists within many applications, such as disease diagnosis, where there may be multiple causes for a particular disease. In this situation, *Fisher discriminant analysis* cannot capture the multimodal characteristics of the clusters. To deal with multimodality, *local-preserving projection* [@lpppaper] preserves the local structure of the data in that it keeps nearby data pairs in the original data space close in the embedding space. As a result, multimodal data could be embedded and its local structure will not be lost.
 
 Later on, a new dimensionality reduction method called *local Fisher discriminant analysis (LFDA)* [@lfdaoriginalpaper] was proposed to combine both advantages of Fisher discriminant analysis and those of local-preserving projection in a way that between-class separability is maximized while within-class separability is minimized and its local structure is preserved. Furthermore, with the help of kernel trick, local Fisher discriminant analysis can also be extended to deal with non-linear dimensionality reduction situations.
 
-*Principal components analysis (PCA)* [@pcapaper] is another popular choice for performing dimension reduction. However, in practice sometimes principal components analysis generates bad principal components that cannot explain a great amount of variance in the original data set. For example, if the original data set has six dimensions and we reduce the dimension into three using *PCA*, the resulted three principal components might not capture some important features and variance in the original data features. Using this result from *PCA* sometimes lead to very bad accuracy for machine learning models due to the poor pre-processing that losses a lot of the essential information in the original data features. On the other hand, metric learning methods such as *LFDA* in particular can, surprisingly, enhance the distinctive characteristics of the original data set and pull data points that have similar characteristics close to each other. However, both *PCA* and *LFDA* have their own shortcomings and a combined approach called *semi-supervised local Fisher discriminant analysis* [@semilfdapaper] mix the supervised and unsupervised approaches to provide a more stable result.
[…]
