### 2019-07-24 be511d3
Fixes #26 and #27: Emphasize statement of need
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
