+++ paper/paper.md
@@ -35,6 +35,6 @@ Later on, a new dimensionality reduction method called *local Fisher discriminan
 
 *Principal components analysis (PCA)* [@pcapaper] is another popular choice for performing dimension reduction. However, in practice sometimes principal components analysis generates bad principal components that cannot explain a great amount of variance in the original dataset. For example, if the original dataset has six dimensions and we reduce the dimension into three using *PCA*, the three principal components might not capture some important characteristics and variance in the original dataset. Using this result from *PCA* sometimes misleads the analysis due to the poor pre-processing that losses a lot of the essential information in the original dataset. On the other hand, metric learning methods such as *LFDA* in particular can, surprisingly, enhance the distinctive characteristics of the original dataset and pull data points that have similar characteristics close to each other. However, both *PCA* and *LFDA* have their own shortcomings and a combined approach called *semi-supervised local Fisher discriminant analysis* [@semilfdapaper] mix the supervised and unsupervised approaches to provide a more stable result.
 
-The `lfda` [@lfdapaper; @lfdapkg] package is an R package for performing *local Fisher discriminant analysis*, *kernel local Fisher discriminant analysis*, and *semi-supervised local Fisher discriminant analysis*. It also provides visualization functions to easily visualize the dimension reduction results by using `rgl` [@rgl] for 3D visualization, `ggfortify` [@rjggfortify; @ggfortify] for 2D visualization in `ggplot2` [@wickham2009ggplot2] style, or `autoplotly` [@autoplotlypkg; @autoplotly2018joss] for interactive visualizations in `plotly` [@plotly] style. `lfda` is also included in `dml` [@dmlpkg; @dmljoss] package together with implementations of other *distance metric learning* algorithms.
+The `lfda` [@lfdapaper; @lfdapkg] package is an R package that provides the implementation for the abovely mentioned methods *local Fisher discriminant analysis*, *kernel local Fisher discriminant analysis*, and *semi-supervised local Fisher discriminant analysis* so researchers could quickly experiment different variations of *local Fisher discriminant analysis* methods in different applications. As of the time of writing, it's the first package with those methods implemented in native R language. It also provides visualization functions to easily visualize the dimension reduction results by using `rgl` [@rgl] for 3D visualization, `ggfortify` [@rjggfortify; @ggfortify] for 2D visualization in `ggplot2` [@wickham2009ggplot2] style, or `autoplotly` [@autoplotlypkg; @autoplotly2018joss] for interactive visualizations in `plotly` [@plotly] style. `lfda` is also included in `dml` [@dmlpkg; @dmljoss] package together with implementations of other *distance metric learning* algorithms.
 
 # References
