### 2019-07-24 04a2fc6
Fixes #25: Add missing DOIs
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
