+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={1},
   number={1},
   pages={1},
-  year={1999}
+  year={1999},
+  doi={10.1109/NNSP.1999.788121}
 }
 
 @InProceedings{lpppaper,
@@ -34,7 +35,8 @@
   issue = {2},
   volume = {8},
   pages={478--489},
-  url = {https://journal.r-project.org/}
+  url = {https://journal.r-project.org/},
+  doi = {10.32614/RJ-2016-060}
 }
 
 @Manual{dmlpkg,
