### 2019-07-24 #1572 @terrytangyuan
@strengejacke Thanks for your review! I have addressed all the issues. Please let me know if there's anything else I need to resolve.
