### 2019-07-24 0b37452
Fixes #30: Fix typo
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
