### 2019-07-25 #1572 @strengejacke
Thanks, all my raised issues have been addressed satisfactorily.

@arfon I recommend publication.
