### 2019-07-25 a13e5c2
Use styler to re-format all the code
Signed-off-by: terrytangyuan <terrytangyuan@gmail.com>
