### 2019-07-26 #1572 @arfon
> Thanks, all my raised issues have been addressed satisfactorily.

> @arfon I also recommend publication.

Wonderful, thanks @strengejacke and @sctyner. Apologies for being absent over the last few days - I've been taking a few days vacation but it looks like things have been progressing nicely here :grin:
