+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: lfda
 Type: Package
 Title: Local Fisher Discriminant Analysis
-Version: 1.1.2
-Date: 2019-07-13
+Version: 1.1.3
+Date: 2019-07-26
 Author: Yuan Tang. Contributions from Wenxuan Li, Nan Xiao and Zachary Deane-Mayer.
 URL: https://github.com/terrytangyuan/lfda
 BugReports: https://github.com/terrytangyuan/lfda/issues
