### 2019-07-26 52aa8b5
Add DOI badge
